@@ -1,6 +1,8 @@
 package com.bodega.api.io;
 
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 import org.hibernate.annotations.Type;
 
 import javax.persistence.*;
@@ -8,19 +10,20 @@ import java.io.Serializable;
 import java.util.List;
 import java.util.UUID;
 
-@Data
+@Getter
+@Setter
 @Entity
 @Table(name = "producto")
 public class ProductEntity implements Serializable {
     @Id
-    @Type(type="org.hibernate.type.PostgresUUIDType")
+    @Type(type="uuid-char")
     public UUID id;
 
     @Column(name = "nombre")
     private String name;
 
     @Column(name = "precio_unitario")
-    private int unitPrice;
+    private double unitPrice;
 
     @Column(name = "cantidad")
     private int quantity;
@@ -28,7 +31,7 @@ public class ProductEntity implements Serializable {
     @Column(name = "imagen")
     private String image;
 
-    @Type(type="org.hibernate.type.PostgresUUIDType")
+    @Type(type="uuid-char")
     @Column(name = "categoria_id", columnDefinition = "uuid")
     private UUID categoryId;
 
